refactor: move scenarios to module
