+++ backend/scenarios.js
@@ -1,0 +1,36 @@
+export const scenarios = [
+  {
+    id: 1,
+    text: `*תרחיש 1*
+חוק המואזין
+בבוקר השיעור אישרה הכנסת, בקריאה ראשונה, הצעת החוק הקוראת על איסור השמעת מואזין
+(מבוסס על הצעת חוק אמיתית). בפתח השיעור, התקיים דיון בנושא, כאשר חלק מהסטודנטים
+הביעו עמדות נחרצות בעד הצעד, מה שעורר תחושות תסכול אצל הסטודנטים המוסלמים. אלה אף נפגעו מעצם קיום הדיון בכיתה.
+כיצד לדעתך המרצה צריך לנהוג בשיעורים הבאים?
+האם עליו לוותר על העיסוק בסוגיה בשל המתיחות? במידה ולא, כיצד היית נוהג.ת?`
+  },
+  {
+    id: 2,
+    text: `*תרחיש 2*
+סטריאוטיפ המוחלש
+בפקולטה ברפואה, הסטודנטים נדרשים לאבחן ולהציע טיפול למטופלים. אחת המתרגלות כתבה
+משימה ובה מקרה בוחן לניתוח עבור הסטודנטים. המקרה מתאר מטופל אתיופי שסובל מאיידס.
+סטודנטים אתיופים שקראו את הדוגמה לא הגישו את המשימה, ונקטו בשביתה איטלקית בקורס
+ועברו אותו בקושי. מנגד, המרצה שלא שמה לב לדוגמה ולתסכול של הסטודנטים, האשימה שהם חסרי מוטיבציה בקורס.
+מה לדעתך אפשר לעשות כדי להימנע ממקרים כאלה בעתיד?`
+  },
+  {
+    id: 3,
+    text: `*תרחיש 3*
+יחסי צוות עכורים במעבדה
+ד"ר לוי מנהלת מעבדת מחקר בכימיה הכוללת שמונה סטודנטים לתואר שני, חצי יהודים וחצי ערבים. בעקבות המתיחות הביטחונית, היחסים בין הקבוצות הדרדרו. סטודנטים שהיו עד לא מכבר ביחסים טובים, הפסיקו לברך זה את זה לשלום והיחסים הצטמצמו להיבטים קורקטיים של העבודה בלבד. ד"ר לוי מרגישה שהיחסים העכורים גם משליכים לרעה על העבודה, משום שבישיבות הצוות הסטודנטים הפסיקו לשתף לעומק, לבקר רעיונות ולהציע רעיונות חדשים.
+כיצד לדעתך ד"ר לוי צריכה לנהוג?`
+  },
+  {
+    id: 4,
+    text: `*תרחיש 4*
+הערות פוגעניות כלפי להט"בים
+בשיעור על פילוסופיה קלאסית במדעי המדינה, הסטודנטים נחשפים לטקסט של אפלטון. אחד הסטודנטים מעיר שסוקרטס תיאר את אפלטון כמי שנמשך לנערים. בתגובה אחד הסטודנטים מעיר "זו סטייה וזה אסור בדת". המרצה לא יודע מה לעשות והוא ממשיך בשיעור מבלי להתייחס לתגובה של הסטודנט. לאחר השיעור, סטודנטית אחרת, שנפגעה מההערה של הסטודנט, כותבת אימייל לראש החוג, בו היא מאשימה את המרצה בחוסר תגובה לגזענות בשיעור ואומרת שהיא לא רוצה להשתתף יותר בשיעור.
+איך לדעתך המרצה צריך לנהוג?`
+  }
+];
